added validations/error handling
