+++ config/docker/realm-config/wda-realm.json
@@ -5,7 +5,7 @@
     "defaultSignatureAlgorithm": "RS256",
     "revokeRefreshToken": false,
     "refreshTokenMaxReuse": 0,
-    "accessTokenLifespan": 300,
+    "accessTokenLifespan": 3600,
     "accessTokenLifespanForImplicitFlow": 900,
     "ssoSessionIdleTimeout": 1800,
     "ssoSessionMaxLifespan": 36000,
@@ -2224,5 +2224,61 @@
     },
     "clientPolicies": {
       "policies": []
-    }
+    },
+    "users": [
+      {
+        "id": "4c973896-5761-41fc-8217-07c5d13a004b",
+        "createdTimestamp": 1505479415590,
+        "username": "admin",
+        "enabled": true,
+        "totp": false,
+        "emailVerified": true,
+        "firstName": "Admin",
+        "lastName": "Administrator",
+        "email": "admin@localhost",
+        "credentials": [
+          {
+            "id": "b860462b-9b02-48ba-9523-d3a8926a917b",
+            "type": "password",
+            "createdDate": 1505479429154,
+            "secretData": "{\"value\":\"4pf9K2jWSCcHC+CwsZP/qidN5pSmDUe6AX6wBerSGdBVKkExay8MWKx+EKmaaObZW6FVsD8vdW/ZsyUFD9gJ1Q==\",\"salt\":\"1/qNkZ5kr77jOMOBPBogGw==\"}",
+            "credentialData": "{\"hashIterations\":27500,\"algorithm\":\"pbkdf2-sha256\"}"
+          }
+        ],
+        "disableableCredentialTypes": [],
+        "requiredActions": [],
+        "realmRoles": ["offline_access", "uma_authorization"],
+        "clientRoles": {
+          "account": ["view-profile", "manage-account"]
+        },
+        "notBefore": 0
+      },
+      {
+        "id": "c4af4e2f-b432-4c3b-8405-cca86cd5b97b",
+        "createdTimestamp": 1505479373742,
+        "username": "user",
+        "enabled": true,
+        "totp": false,
+        "emailVerified": true,
+        "firstName": "",
+        "lastName": "User",
+        "email": "user@localhost",
+        "credentials": [
+          {
+            "id": "7821832b-1e82-45a2-b8d3-f1a6ad909e64",
+            "type": "password",
+            "createdDate": 1505479392766,
+            "secretData": "{\"value\":\"MbKsMgWPnZyImih8s4SaoCSCq+XIY/c6S9F93sXEidHF1TjPWxCqMkec0+o3860CMLXHt3az61cIJOWI0FW9aw==\",\"salt\":\"fmpBI1r8R1u75hDLMUlwBw==\"}",
+            "credentialData": "{\"hashIterations\":27500,\"algorithm\":\"pbkdf2-sha256\"}"
+          }
+        ],
+        "disableableCredentialTypes": [],
+        "requiredActions": [],
+        "realmRoles": ["offline_access", "uma_authorization"],
+        "clientRoles": {
+          "account": ["view-profile", "manage-account"]
+        },
+        "notBefore": 0
+      }
+    ]
   }
